Add members `priority` and `score` to struct `process`
`priority` can be used by the scheduler when choosing which thread should run
next. `priority` is an enumerator that can have values of LOW, NORMAL or HIGH.
The default priority of each process is NORMAL.

`score` and `priority` can be used together to make modestly nuanced decisions
about scheduling. The end goal is this:

    1. Each time `Schedule()` is called, bump the score of each process. The
       more the score is bumped.
    2. Select the process with the highest score, reset the score to 0, and run
       it. If multiple processes have the same score, run the first process
       after the currently selected process.