Update roundbuffer indexing functions
Delete `set_idle()` and add `set_current()`. The former was a hardcoded hack,
and the latter is a bit more flexible.

The ultimate aim here is to change the roundbuffer functions so that "Only
`add_process()` and `set_current()` affect the index." This will allow
`Schedule()` to more easily walk through the roundbuffer without worrying about
accidentally affecting the index.